Retain speaker ON setting when SCO is disconnected
Use case:
1. Connected BT headset.
2. Receive call.
3. Transfer call audio to Phone speaker.
4. Transfer call audio to BT headset.
5. Transfer call audio to Phone.

Failure:
After step 5, call audio is on earpiece instead on
speaker as user selected previously.

Root cause:
When SCO is disconnected, BT is not checking if current
audio device selected is speaker. AudioManager is
routing call audio to default audio end point.

Fix:
When sco is disconnected, check if the preferred
device to use is speaker, enable the  same in
audio manager.

Change-Id: Icc3f20c6490eec0afc43b5292294b31770f69801